test/django_school/requirements.txt
Django==2.0.1
django-crispy-forms==1.7.0
django-phone-field==1.8.1
Pillow==7.2.0
pytz==2017.3
